Fixed naming when setting action all caps
It was previously referring it to "lowercase action", however, this isn't the case. What it actually does is disable the "text all caps" setting. So it is choosing between forcing text all caps (default above API 14) and using the case from the original string.

Fixed this naming inconsistency in SnackbarBuilder and SnackbarWrapper. This involved removing some individual methods and just using setter methods that take the allCaps boolean. This approach is more specific and generally cleaner in the API anyway.